docs: restructure and improve Go Bank Transfer Quest documentation
- Updated main README to simplify Module 3 references
- Revamped internal/bank/README.md for layered architecture overview
- Interconnected Domain, Repository, Service, and API READMEs
- Clarified Quest instructions and formatted Definitions of Done with Makefile commands
- Added OpenAPI design instructions

+++ internal/bank/domain/transfer.go
@@ -6,5 +6,5 @@ package domain
 type Transfer struct {
 	FromAccountID string
 	ToAccountID   string
-	Amount        float64
+	Amount        int64 // in cents
 }
