+++ migration/001_create_accounts.sql
@@ -4,7 +4,7 @@
 CREATE TABLE IF NOT EXISTS accounts (
     id          VARCHAR(64)   NOT NULL PRIMARY KEY,
     owner       VARCHAR(255)  NOT NULL,
-    balance     NUMERIC(15,4) NOT NULL DEFAULT 0,
+    balance     BIGINT NOT NULL DEFAULT 0,
     status      VARCHAR(16)   NOT NULL DEFAULT 'OPEN',
     created_at  TIMESTAMPTZ   NOT NULL DEFAULT NOW(),
     updated_at  TIMESTAMPTZ   NOT NULL DEFAULT NOW()
